# V: improve arrays allocations