This PR converts `top5` and `top_posts` arrays to fixed ones since their size is always the same. This is a performance improvement because fixed arrays are always allocated on the stack.

Also, this PR changes an algorithm that calculates the top 5 posts to be equal to the F# version.