@jinyus 2023-10-18
Main branch:
```
51.53 ms 
756.98 ms 
6.67 s
```

This PR:
```
Vlang:

        Benchmark 1: ./related
        Processing time (w/o IO): 48.753ms
        Processing time (w/o IO): 49.127ms
        Processing time (w/o IO): 48.993ms
        Processing time (w/o IO): 48.993ms
        Processing time (w/o IO): 48.472ms
        Processing time (w/o IO): 48.812ms
        Processing time (w/o IO): 48.998ms
        Processing time (w/o IO): 48.891ms
        Processing time (w/o IO): 48.719ms
        Processing time (w/o IO): 48.619ms
        Processing time (w/o IO): 49.009ms
        Processing time (w/o IO): 48.875ms
        Processing time (w/o IO): 48.740ms
          Time (mean ± σ):     370.4 ms ±   3.0 ms    [User: 344.7 ms, System: 25.6 ms]
          Range (min … max):   366.1 ms … 374.3 ms    10 runs
         
Vlang:

        Benchmark 1: ./related
        Processing time (w/o IO): 746.271ms
        Processing time (w/o IO): 739.842ms
        Processing time (w/o IO): 740.613ms
          Time (mean ± σ):      8.698 s ±  0.001 s    [User: 8.571 s, System: 0.124 s]
          Range (min … max):    8.697 s …  8.698 s    2 runs
         
Vlang:

        Benchmark 1: ./related
        Processing time (w/o IO): 6.597s
        Processing time (w/o IO): 6.596s
        Processing time (w/o IO): 6.593s
          Time (mean ± σ):     77.773 s ±  1.193 s    [User: 77.375 s, System: 0.385 s]
          Range (min … max):   76.930 s … 78.617 s    2 runs

```